R7-JAXRS integration test
The currently available version of aries-jax-rs-whiteboard is having classloading issues if used with pax-web and scr.

It is not clear if this is a bug in aries-jax-rs-whiteboard or pax-web or scr.

The provided sample and integration test are meant to reproduce the classloading-issue.

Under the hood: aries-jax-rs-whiteboard ships together with all CXF dependencies included. The RuntimeDelegate from the jax-rs-api is trying to locate an implementation via /META-INF/services which is provided by the included cxf-frontend-jaxrs so the classloader should be able to locate the file...it doesnt